requirements.txt
apache-airflow==1.10.10
#user conda install pyarrow
#pyarrow==0.17.1
oss2==2.11.0
mysqlclient==1.3.14
kubernetes==11.0.0
cryptography==2.9.2
pyhive==0.6.2
sasl==0.2.1
thrift-sasl==0.4.2
hmsclient==0.1.1
PyMySQL==0.9.3